chore: type收敛
